Use an array cache to store properties and values
Objects that are used as maps can create internal values in the vm when
looking up if the object has a key of a given value. Caching that
detail in the array and doing a fast scan over recently looked up
values can offer some benefit reducing memory thrashing and gaining a
little performance.